# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CPA Intern Radar API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Firm {
  id: number;
  name: string;
  rank: number;
  hqCity: string;
  hqState: string;
  websiteUrl: string;
  careersUrl: string;
  /** ATS provider: workday, greenhouse, lever, icims, smartrecruiters, ashby, jobvite, custom, unknown */
  atsType: string;
  /** @nullable */
  atsUrl?: string | null;
  /** @nullable */
  lastChecked?: string | null;
  jobsCount?: number;
  createdAt: string;
}

export interface FirmInput {
  name: string;
  rank: number;
  hqCity: string;
  hqState: string;
  websiteUrl: string;
  careersUrl: string;
  atsType?: string;
  atsUrl?: string;
}

export interface FirmUpdate {
  name?: string;
  rank?: number;
  hqCity?: string;
  hqState?: string;
  websiteUrl?: string;
  careersUrl?: string;
  atsType?: string;
  /** @nullable */
  atsUrl?: string | null;
}

export interface AtsDetectionResult {
  firmId: number;
  atsType: string;
  /** @nullable */
  detectedUrl: string | null;
  success: boolean;
  /** @nullable */
  message?: string | null;
}

export interface Job {
  id: number;
  firmId: number;
  firmName: string;
  title: string;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  applyUrl?: string | null;
  /** @nullable */
  term?: string | null;
  atsSource: string;
  isActive: boolean;
  firstSeen: string;
  lastSeen: string;
}

export interface Digest {
  date: string;
  newToday: Job[];
  stillOpen: Job[];
  totalActive: number;
}

export interface ScrapeResult {
  startedAt: string;
  firmsProcessed: number;
  jobsFound: number;
  jobsNew: number;
  errors: string[];
}

export interface FirmScrapeResult {
  firmId: number;
  firmName: string;
  atsType: string;
  jobsFound: number;
  jobsNew: number;
  success: boolean;
  /** @nullable */
  error?: string | null;
}

export interface AtsBucket {
  atsType: string;
  count: number;
}

export interface Stats {
  totalFirms: number;
  totalActiveJobs: number;
  newJobsToday: number;
  firmsScannedToday: number;
  firmsByAts: AtsBucket[];
  recentJobs: Job[];
}

export type ListFirmsParams = {
ats_type?: string;
search?: string;
};

export type ListJobsParams = {
firm_id?: number;
is_active?: boolean;
search?: string;
term?: string;
ats_source?: string;
};