# lib/api-zod/src/generated/types/atsDetectionResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CPA Intern Radar API
 * OpenAPI spec version: 0.1.0
 */

export interface AtsDetectionResult {
  firmId: number;
  atsType: string;
  /** @nullable */
  detectedUrl: string | null;
  success: boolean;
  /** @nullable */
  message?: string | null;
}