# lib/api-zod/src/generated/types/digest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CPA Intern Radar API
 * OpenAPI spec version: 0.1.0
 */
import type { Job } from './job';

export interface Digest {
  date: string;
  newToday: Job[];
  stillOpen: Job[];
  totalActive: number;
}