# lib/api-zod/src/generated/types/firm.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CPA Intern Radar API
 * OpenAPI spec version: 0.1.0
 */

export interface Firm {
  id: number;
  name: string;
  rank: number;
  hqCity: string;
  hqState: string;
  websiteUrl: string;
  careersUrl: string;
  /** ATS provider: workday, greenhouse, lever, icims, smartrecruiters, ashby, jobvite, custom, unknown */
  atsType: string;
  /** @nullable */
  atsUrl?: string | null;
  /** @nullable */
  lastChecked?: string | null;
  jobsCount?: number;
  createdAt: string;
}