# lib/api-zod/src/generated/types/firmInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CPA Intern Radar API
 * OpenAPI spec version: 0.1.0
 */

export interface FirmInput {
  name: string;
  rank: number;
  hqCity: string;
  hqState: string;
  websiteUrl: string;
  careersUrl: string;
  atsType?: string;
  atsUrl?: string;
}