# lib/api-zod/src/generated/types/firmScrapeResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CPA Intern Radar API
 * OpenAPI spec version: 0.1.0
 */

export interface FirmScrapeResult {
  firmId: number;
  firmName: string;
  atsType: string;
  jobsFound: number;
  jobsNew: number;
  success: boolean;
  /** @nullable */
  error?: string | null;
}