# lib/api-zod/src/generated/types/job.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CPA Intern Radar API
 * OpenAPI spec version: 0.1.0
 */

export interface Job {
  id: number;
  firmId: number;
  firmName: string;
  title: string;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  applyUrl?: string | null;
  /** @nullable */
  term?: string | null;
  atsSource: string;
  isActive: boolean;
  firstSeen: string;
  lastSeen: string;
}