# lib/api-zod/src/generated/types/scrapeResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CPA Intern Radar API
 * OpenAPI spec version: 0.1.0
 */

export interface ScrapeResult {
  startedAt: string;
  firmsProcessed: number;
  jobsFound: number;
  jobsNew: number;
  errors: string[];
}