# lib/api-zod/src/generated/types/stats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CPA Intern Radar API
 * OpenAPI spec version: 0.1.0
 */
import type { AtsBucket } from './atsBucket';
import type { Job } from './job';

export interface Stats {
  totalFirms: number;
  totalActiveJobs: number;
  newJobsToday: number;
  firmsScannedToday: number;
  firmsByAts: AtsBucket[];
  recentJobs: Job[];
}